Add a test for https://github.com/dotnet/roslyn/issues/33080
We never actually had any tests that tested Code Model and our project
system shims at the same time. To fix this required pulling
RoslynVisualStudioWorkspace.GetFileCodeModel down to
VisualStudioWorkspaceImpl, and fixing up some mocks we have to be
(every so) slightly more accurate.
