+++ src/VisualStudio/Core/Def/Implementation/CodeModel/IProjectCodeModelFactory.cs
@@ -8,5 +8,6 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.CodeModel
     internal interface IProjectCodeModelFactory
     {
         IProjectCodeModel CreateProjectCodeModel(ProjectId id, ICodeModelInstanceFactory codeModelInstanceFactory);
+        EnvDTE.FileCodeModel GetOrCreateFileCodeModel(ProjectId id, string filePath);
     }
 }
